Move attribute parser to a separate module
